requirements.txt
django
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
python-decouple
reportlab
django-filter
